chore(memory): address final review minors — stale refs, accept() docstring, task fall-through test, helper module rename
